+++ src/main/java/com/jica/sdg/repository/EntryNsaBudgetRepository.java
@@ -9,7 +9,7 @@ import javax.transaction.Transactional;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.repository.query.Param;
 
-public interface EntryNsaBudgetRepository extends CrudRepository<EntryNsaBudget, Long> {
+public interface EntryNsaBudgetRepository extends CrudRepository<EntryNsaBudget, Integer> {
 
     
 }
